requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dnspython==2.7.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-PyMongo==2.3.0
frozendict==2.4.6
html5lib==1.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
multitasking==0.0.11
numpy==2.0.2
pandas==2.2.3
peewee==3.17.7
platformdirs==4.3.6
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-futures==1.0.1
six==1.16.0
soupsieve==2.6
tqdm==4.66.6
tzdata==2024.2
urllib3==2.2.3
webencodings==0.5.1
Werkzeug==3.1.1
yfinance==0.2.48
zipp==3.20.2
